[CCUBE-1492] Reorganise pill components
