--- src/components/common/pills/pills.styles.tsx
@@ -1,6 +1,0 @@
-import { ButtonWithIcon } from "@lifesg/react-design-system/button-with-icon";
-import styled from "styled-components";
-
-export const AddMultiEntryButton = styled(ButtonWithIcon.Default)`
-    margin: 0 0 2rem;
-`;
