+++ src/components/common/pills/pills.tsx
@@ -1,128 +1,102 @@
-import {
-    DndContext,
-    DragEndEvent,
-    KeyboardSensor,
-    PointerSensor,
-    closestCenter,
-    useSensor,
-    useSensors,
-} from "@dnd-kit/core";
-import {
-    SortableContext,
-    sortableKeyboardCoordinates,
-    verticalListSortingStrategy,
-} from "@dnd-kit/sortable";
-import { Form } from "@lifesg/react-design-system/form";
-import { PlusIcon } from "@lifesg/react-icons/plus";
-import { Controller, useFieldArray, useFormContext } from "react-hook-form";
+import { Controller, useFormContext } from "react-hook-form";
 import { TogglePair } from "src/components/common/toggle-pair/toggle-pair";
-import { ITextareaFieldAttributes } from "src/context-providers";
+import { MandatoryFieldBox } from "src/components/element-editor/basic-details/basic-details.styles";
+import {
+    ITextarea,
+    ITextareaFieldAttributes,
+    useBuilder,
+} from "src/context-providers";
 import { BottomPositionIcon } from "../icons/bottom-position-icon";
 import { TopPositionIcon } from "../icons/top-position-icon";
-import { PillItem } from "./pill-items";
-import { AddMultiEntryButton } from "./pills.styles";
+import { PillItems } from "./pill-items";
 
 export const Pills = () => {
     // =========================================================================
     // CONST, STATE, REF
     // =========================================================================
-    const { control } = useFormContext<ITextareaFieldAttributes>();
-    const { fields, append, remove, move } = useFieldArray({
-        control,
-        name: "pillItems",
-        shouldUnregister: true,
-    });
+    const { control, unregister, setValue, watch } =
+        useFormContext<ITextareaFieldAttributes>();
 
-    const sensors = useSensors(
-        useSensor(PointerSensor),
-        useSensor(KeyboardSensor, {
-            coordinateGetter: sortableKeyboardCoordinates,
-        })
-    );
-
-    const disableDeleteButton = fields.length <= 2;
-    // =========================================================================
-    // EVENT HANDLERS
-    // =========================================================================
-    const handleDragEnd = (event: DragEndEvent) => {
-        const { active, over } = event;
-
-        if (over && active.id !== over.id) {
-            const oldIndex = fields.findIndex((item) => item.id === active.id);
-            const newIndex = fields.findIndex((item) => item.id === over.id);
-            move(oldIndex, newIndex);
-        }
-    };
-
-    const handleAddButtonClick = () => {
-        append({ content: "" });
-    };
-
-    const handleDeleteButtonClick = (index: number) => {
-        remove(index);
-    };
+    const { focusedElement } = useBuilder();
+    const element = focusedElement?.element as ITextarea;
 
     // =============================================================================
     // RENDER FUNCTIONS
     // =============================================================================
-    const renderPillItems = () =>
-        fields.map((item, index) => (
-            <PillItem
-                key={item.id}
-                item={item}
-                index={index}
-                onDelete={() => handleDeleteButtonClick(index)}
-                disableDelete={disableDeleteButton}
-            />
-        ));
+
+    const renderPillsToggle = () => (
+        <Controller
+            name="pills"
+            control={control}
+            defaultValue={false}
+            render={({ field }) => (
+                <TogglePair
+                    label={{
+                        mainLabel: "Pills",
+                        subLabel:
+                            "This allows a list of selectable short texts to display in pill form. This helps participant to fill up text area fast with less typing.",
+                    }}
+                    value={field.value}
+                    onChange={(value) => {
+                        if (!value) {
+                            /** Without this, the form remains dirtied when toggling Pills to true and back to false */
+                            unregister(["pillItems", "pillPosition"]);
+                        } else {
+                            setValue("pillItems", [
+                                {
+                                    content: "",
+                                },
+                                {
+                                    content: "",
+                                },
+                            ]);
+                            setValue("pillPosition", "top");
+                        }
+                        field.onChange(value);
+                    }}
+                />
+            )}
+            shouldUnregister={true}
+        />
+    );
+
+    const renderListPositionToggle = () => (
+        <Controller
+            name="pillPosition"
+            control={control}
+            render={({ field }) => (
+                <TogglePair
+                    label={{
+                        mainLabel: "List position",
+                        subLabel:
+                            "This displays pills on top or bottom of the text area.",
+                    }}
+                    value={field.value === "top"}
+                    options={[
+                        { icon: <TopPositionIcon />, title: "Top" },
+                        {
+                            icon: <BottomPositionIcon />,
+                            title: "Bottom",
+                        },
+                    ]}
+                    onChange={(value) =>
+                        field.onChange(value ? "top" : "bottom")
+                    }
+                />
+            )}
+            shouldUnregister
+        />
+    );
 
     return (
-        <DndContext
-            sensors={sensors}
-            collisionDetection={closestCenter}
-            onDragEnd={handleDragEnd}
-        >
-            <SortableContext
-                items={fields}
-                strategy={verticalListSortingStrategy}
-            >
-                <Form.Label>Pill items</Form.Label>
-                {renderPillItems()}
-                <AddMultiEntryButton
-                    icon={<PlusIcon />}
-                    styleType="light"
-                    onClick={handleAddButtonClick}
-                    role="button"
-                    type="button"
-                >
-                    Add option
-                </AddMultiEntryButton>
-            </SortableContext>
-            <Controller
-                name="pillPosition"
-                control={control}
-                render={({ field }) => (
-                    <TogglePair
-                        label={{
-                            mainLabel: "List position",
-                            subLabel:
-                                "This displays pills on top or bottom of the text area.",
-                        }}
-                        value={field.value === "top"}
-                        options={[
-                            { icon: <TopPositionIcon />, title: "Top" },
-                            {
-                                icon: <BottomPositionIcon />,
-                                title: "Bottom",
-                            },
-                        ]}
-                        onChange={(value) =>
-                            field.onChange(value ? "top" : "bottom")
-                        }
-                    />
-                )}
-                shouldUnregister
-            />
-        </DndContext>
+        <MandatoryFieldBox>
+            {renderPillsToggle()}
+            {watch("pills", element.pills) && (
+                <>
+                    <PillItems />
+                    {renderListPositionToggle()}
+                </>
+            )}
+        </MandatoryFieldBox>
     );
 };
